Init Worker in order to improve performance
As Jen pointed out (#18), big split files (like Any% NMMS with ~4000
runs) can be extremely slow to process (to the point where browser will
freeze for a while before the website displays properly. This is an
attempt at improving that, by offloading some of the processing to a
Web Worker. For now it's only implemented for a few functions, but it
already improves loading by *a lot* (in particular there is no
noticeable freeze any more). The tradeoff is that the code is much less
readable and significantly more complex in general (since all processing
that goes through the worker is highly asynchronous, and has to go
through a (de)serialisation process on both side each time some
processing is requested)