Fairly big visual re-design
Mostly:
- Change background from a dull grey to dark blue
- Make the logo shrink when a splitfile is loaded
- Organise cards into 4 tabs. Unfortunately the options don't really fit
in either tab so for now they're still a "global" thing sitting on top
of the tabs, but hopefully we can improve that at some point
- Update Cypress tests to match the new layout

Also fix the "game time / real time" switch. It stopped working for two
reasons which are now fixed:
- For some reason selectTime ignored it, I must have removed this logic
when trying to optimize performance but obviously it made no sense
- The web worker doesn't share data with the main application except for
values that are explicitely passed using messages. Therefore selectTime
inside the worker was trying to access useRealTime from the global
state, but that didn't map to the correct data (it refered to a "local"
version of the store). Therefore a new operation type was added to
update the Worker's store whenever the main app updates its store
(currently this logic is done inside SplitsDisplay but that doesn't
really matter, it could be done in any component that's always loaded)