Better subsplits management (episode 1)
Compute an improved "intermediate representation" of splits where
subsplits are properly stored inside their "bigger" split so that we
can display them as a "tree" in a prettier (and more understandable)
way.

Computing this representation is done inside the web worker as it is
costly and probably not very optimized (currently).

We still need to use this pretty representation is the comparison tab
where similar info is displayed.